Change plugin to integrate into nu structure and build system
