+++ src/plugins/nu_plugin_stable_shuffle.rs
@@ -1,0 +1,6 @@
+use nu_plugin::serve_plugin;
+use nu_plugin_shuffle::Shuffle;
+
+fn main() {
+    serve_plugin(&mut Shuffle::new());
+}
